+++ src/views/TextScroll.vue
@@ -6,17 +6,20 @@ const scrollText = ref<any[]>([
     link: 'https://blog.csdn.net/Dandrose?type=blog'
   },
   {
-    title: '首次出版于1951年'
+    title: '首次出版于1951年',
+    link: 'https://blog.csdn.net/Dandrose?type=blog'
   },
   {
     title:
       '塞林格将故事的起止局限于16岁的中学生霍尔顿·考尔菲德从离开学校到纽约游荡的三天时间内,塞林格将故事的起止局限于16岁的中学生霍尔顿·考尔菲德从离开学校到纽约游荡的三天时间内'
   },
   {
-    title: '并借鉴了意识流天马行空的写作方法，充分探索了一个十几岁少年的内心世界'
+    title: '并借鉴了意识流天马行空的写作方法，充分探索了一个十几岁少年的内心世界',
+    link: 'https://blog.csdn.net/Dandrose?type=blog'
   },
   {
-    title: '愤怒与焦虑是此书的两大主题，主人公的经历和思想在青少年中引起强烈共鸣'
+    title: '愤怒与焦虑是此书的两大主题，主人公的经历和思想在青少年中引起强烈共鸣',
+    link: 'https://blog.csdn.net/Dandrose?type=blog'
   }
 ])
 const singleText = {
@@ -26,9 +29,9 @@ const singleText = {
 const textScroll = ref()
 const disabled = ref(true)
 const vertical = ref(false)
-function onClick(text: any) {
+function onClick(text: string) {
   // 获取点击的标题
-  console.log('text:', text)
+  console.log('text', text)
 }
 function handleStart() {
   textScroll.value.start()
@@ -62,8 +65,6 @@ const state = reactive({
     <h1>{{ $route.name }} {{ $route.meta.title }}</h1>
     <h2 class="mt30 mb10">水平文字滚动</h2>
     <TextScroll :scrollText="scrollText" @click="onClick" />
-    <h2 class="mt30 mb10">自定义滚动速度</h2>
-    <TextScroll :scrollText="scrollText" :step="2" @click="onClick" />
     <h2 class="mt30 mb10">单条文字滚动</h2>
     <TextScroll
       :scrollText="singleText"
@@ -80,6 +81,8 @@ const state = reactive({
       vertical
       @click="onClick"
     />
+    <h2 class="mt30 mb10">自定义链接悬浮样式</h2>
+    <TextScroll :scrollText="scrollText" link-hover-color="#ff6900" @click="onClick" />
     <h2 class="mt30 mb10">自定义样式</h2>
     <TextScroll
       :scrollText="scrollText"
@@ -104,6 +107,8 @@ const state = reactive({
       </Space>
       <TextScroll :vertical="vertical" ref="textScroll" :scrollText="scrollText" @click="onClick" />
     </Flex>
+    <h2 class="mt30 mb10">自定义滚动速度</h2>
+    <TextScroll :scrollText="scrollText" :step="2" @click="onClick" />
     <h2 class="mt30 mb10">文字滚动配置器</h2>
     <Row :gutter="[24, 12]">
       <Col :span="6">